requirements.txt
streamlit
pandas
pillow
pytesseract
pdf2image
passlib
python-dotenv